SDL3 + SDL_Gpu (#193)
* CF builds on SDL3 and can run (no rendering)

* Created an SDL_GpuShader!!!

* Remove deleted file references

* Clear screen works

* WIP not working at all

* WIP (super broken still)

* WIP totes broke

* WIP (broken)

* Rendering working, still some bugs

* pipeline caching

* Tweaks

* imgui not crashing

* Fairly stable

* Explicitly cast returned pointer in cute_symbol to void* (#186)

* Fix SIMD compilation for cute_sound on gcc (#187)

* Enable OBJC on Apple platforms (#185)

* Sync to latest SDL_Gpu

* Fix window flags, resizing seems to work

* Remove erronous extra canvas clears

* Remove extraneous blit code

* scissor height not being used (#189)

* Attempt macOS NEON port for cute_sound.h

* Attempt to get MinGW correct D3D12 headers

* Number typecast fixes for macOS build

* Remove chocolatey winsdk install thing

* Disable MinGW since it has out of date DX12 headers

* Add tasks_running count and wait for it to reach 0 before returning from cute_threadpool_kick_and_wait (#190)

* Remove extra tokens (#191)

* Resize vertex buffer if necessary in cf_mesh_update_vertex_data (#192)

* Resize vertex buffer if necessary in cf_mesh_update_vertex_data

* Double the size when resizing vertex buffers

* sync to latest SDL_Gpu

* set vsync flags upon window claim

* remove custom blit and use SDL_GpuBlit

* Clear up flags used for testing in the sample code

* Remove tint and ensure blit stretches to screen

* docs pass

* Stencil outline sample + docs pass

* Tidy

* Actually tidy

* Workaround D3D11 bugs

* Fix memory leaks in tests

* Clear up mem leaks

* Try and disable s2n from building tests

---------

+++ include/cute_input.h
@@ -659,7 +659,7 @@ CF_API void CF_CALL cf_register_key_callback(void (*key_callback)(CF_KeyButton k
  * @remarks  (0, 0) is the top-left of the screen, y-downards.
  * @related  CF_MouseButton cf_mouse_down cf_mouse_x cf_mouse_y
  */
-CF_API int CF_CALL cf_mouse_x();
+CF_API float CF_CALL cf_mouse_x();
 
 /**
  * @function cf_mouse_y
@@ -668,7 +668,7 @@ CF_API int CF_CALL cf_mouse_x();
  * @remarks  (0, 0) is the top-left of the screen, y-downwards.
  * @related  CF_MouseButton cf_mouse_down cf_mouse_x cf_mouse_y
  */
-CF_API int CF_CALL cf_mouse_y();
+CF_API float CF_CALL cf_mouse_y();
 
 /**
  * @function cf_mouse_down
@@ -700,7 +700,7 @@ CF_API bool CF_CALL cf_mouse_just_released(CF_MouseButton button);
  * @brief    Returns a signed integer representing by how much the mouse wheel was rotated.
  * @related  CF_MouseButton cf_mouse_down cf_mouse_x cf_mouse_y cf_mouse_just_pressed cf_mouse_just_released cf_mouse_wheel_motion cf_mouse_double_click_held cf_mouse_double_clicked
  */
-CF_API int CF_CALL cf_mouse_wheel_motion();
+CF_API float CF_CALL cf_mouse_wheel_motion();
 
 /**
  * @function cf_mouse_double_click_held
@@ -983,13 +983,13 @@ CF_INLINE bool key_alt() { return cf_key_alt(); }
 CF_INLINE bool key_gui() { return cf_key_gui(); }
 CF_INLINE void clear_key_states() { cf_clear_key_states(); }
 
-CF_INLINE int mouse_x() { return cf_mouse_x(); }
-CF_INLINE int mouse_y() { return cf_mouse_y(); }
+CF_INLINE float mouse_x() { return cf_mouse_x(); }
+CF_INLINE float mouse_y() { return cf_mouse_y(); }
 
 CF_INLINE bool mouse_down(MouseButton button) { return cf_mouse_down(button); }
 CF_INLINE bool mouse_just_pressed(MouseButton button) { return cf_mouse_just_pressed(button); }
 CF_INLINE bool mouse_just_released(MouseButton button) { return cf_mouse_just_released(button); }
-CF_INLINE int mouse_wheel_motion() { return cf_mouse_wheel_motion(); }
+CF_INLINE float mouse_wheel_motion() { return cf_mouse_wheel_motion(); }
 CF_INLINE bool mouse_double_click_held(MouseButton button) { return cf_mouse_double_click_held(button); }
 CF_INLINE bool mouse_double_clicked(MouseButton button) { return cf_mouse_double_clicked(button); }
 CF_INLINE void mouse_hide(bool true_to_hide) { cf_mouse_hide(true_to_hide); }
